+++ src/EtsyApiSharp/Models/Pong.cs
@@ -2,12 +2,16 @@ using System.Text.Json.Serialization;
 
 namespace EtsyApiSharp.Models
 {
-    //A confirmation that the current application has access to the Open API
+    /// <summary>
+    /// A confirmation that the current application has access to the Open API
+    /// </summary>
     public class Pong
     {
+        /// <summary>
+        /// The authenticated application's ID
+        /// </summary>
         [JsonPropertyName("application_id")]
-        public int ApplicationId { get; set; }
-
+        public long ApplicationId { get; set; }
 
     }
 }
