+++ src/EtsyApiSharp/Models/Shops/ShopProductionPartner.cs
@@ -2,20 +2,28 @@ using System.Text.Json.Serialization;
 
 namespace EtsyApiSharp.Models
 {
-    //Represents a description of a shop production partner.
+    /// <summary>
+    /// Represents a description of a shop production partner.
+    /// </summary>
     public class ShopProductionPartner
     {
+        /// <summary>
+        /// The numeric ID of a production partner.
+        /// </summary>
         [JsonPropertyName("production_partner_id")]
-        public int ProductionPartnerId { get; set; }
+        public long ProductionPartnerId { get; set; }
 
-
+        /// <summary>
+        /// The name or title of the production partner.
+        /// </summary>
         [JsonPropertyName("partner_name")]
         public string PartnerName { get; set; }
 
-
+        /// <summary>
+        /// A string representing the production partner location.
+        /// </summary>
         [JsonPropertyName("location")]
         public string Location { get; set; }
-
 
     }
 }
